Added references and investigated data
Where I can tell that the drone settled into a waypoint based on the yaw signaling, the filtered tilt does look like it increases with altitude, as you might expect if the drone was encountering stronger and stronger winds.

Today's weather report suggests I actually during the windiest part of the day, so we should try to fly first thing tomorrow morning when it's forecast to be calm.